pyproject.toml
[tool.poetry]
name = "identity-backend"
version = "0.1.0"
description = ""
authors = ["cxzlw <wsc_zds@qq.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "<4.0, >= 3.9"
fastapi = "^0.115.2"
pyjwt = "^2.9.0"
sqlalchemy = "^2.0.36"
aiosqlite = "^0.20.0"
argon2-cffi = "^23.1.0"
black = "^24.10.0"
uvicorn = "^0.32.0"
python-multipart = "^0.0.12"
pydantic-settings = "^2.6.0"
rich = "^13.9.2"
pyyaml = "^6.0.2"
ruff = "^0.7.0"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "B",
    "SIM",
    "I",
    "UP",
    "RUF",
    "D",
]

ignore = [
    "B008",
    "W191",
    "E117",
    "E111",
    "E114",
    "D206",
    "D300",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "COM812",
    "COM819",
    "ISC001",
    "ISC002",
    "E501",
    "E701",
    "E203",
    "D202",
    "D203",
    "D213",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
